fix: use 6 River ESLint configuration
The majority of the code changes were made with `eslint --fix`, so they
should have zero impact on the functioning of the code. The majority of
the remaining problems after that were violations of the `eqeqeq` rule,
which is difficult to change without a strong unit test suite, so
comments were used to disable the rules for most of those lines.